requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2
opencv-python==4.8.0.74
Pillow==10.0.0
scikit-learn==1.3.0
scipy==1.11.1
statsmodels==0.14.0
prophet==1.1.4
google-generativeai==0.3.0
python-dotenv==1.0.0
fpdf==1.7.2
reportlab==4.0.4
xlsxwriter==3.1.2
openpyxl==3.1.2